Organized and cleaned up files structure, separating app.js into two modules for the game controller and the board functions, changed color palette
