+++ css/style.css
@@ -7,19 +7,13 @@
 }
 
 body {
-    background-color: #191970;
-}
-
-
-header {
-    font-family: "VT323", monospace;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
+    font-family: Arial, sans-serif;
+    text-align: center;
+    background-color: #F4F4F4;
 }
 
 h1 {
+    font-family: "VT323", monospace;
     font-size: 6vw;
     color: #00FF00;
 }
@@ -28,8 +22,7 @@ h2 {
     font-size: 3rem;
 }
 
-p,
-button {
+p {
     font-family: "Roboto", Verdana, Geneva, Tahoma, sans-serif;
     font-size: 1.2rem;
 }
@@ -41,26 +34,26 @@ main {
 }
 
 button {
-    padding: 12px;
-    background-color: blue;
-    color: wheat;
-    border-radius: 10px;
+    background: linear-gradient(135deg, #ff7eb3, #ff758c);
+    border: none;
+    padding: 10px 20px;
+    margin: 5px;
+    font-size: 20px;
+    color: white;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: .3s
 }
 
 button:hover {
-    background-color: darkblue;
-}
-
-.row {
-    display: flex;
-    justify-content: center;
+    filter: brightness(1.2)
 }
 
 .tile {
     height: 100%;
     aspect-ratio: 1;
-    background-color: gray;
-    border: 1px solid black;
+    background-color: var(--tile-color, #CCC);
+    border: 1px solid #999;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -75,6 +68,49 @@ button:hover {
     background-color: red;
 }
 
+.bg-color-0 {
+    --tile-color: darkgray
+}
+
+.bg-color-1 {
+    --tile-color: blue
+}
+
+.bg-color-2 {
+    --tile-color: green
+}
+
+.bg-color-3 {
+    --tile-color: red
+}
+
+.bg-color-4 {
+    --tile-color: purple
+}
+
+.bg-color-5 {
+    --tile-color: maroon
+}
+
+.bg-color-6 {
+    --tile-color: turquoise
+}
+
+.bg-color-7 {
+    --tile-color: black
+}
+
+.bg-color-8 {
+    --tile-color: gray
+}
+
+#game-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 20px;
+}
+
 #game-info {
     font-family: Verdana, Geneva, Tahoma, sans-serif;
     display: none;
@@ -85,7 +121,7 @@ button:hover {
 }
 
 #game-board {
-    display: none;
+    gap: 2px;
     flex-direction: column;
     justify-content: center;
     align-items: center;
@@ -119,13 +155,15 @@ button:hover {
     border-radius: 10px;
     padding: 20px;
     text-align: center;
+    font-family: "VT323", monospace;
+    display: flex;
 }
 
-#start-game {
-    padding: 10px 20px;
-    color: white;
-    border-radius: 5px;
-    cursor: pointer;
+#game-board{
+    grid-template-columns: repeat(var(--board-size), 1fr);
+    grid-template-rows: repeat(var(--board-size), 1fr);
+    gap: 2px;
+
 }
 
 #level-selector {
@@ -171,18 +209,8 @@ button:hover {
     font-size: 8rem;
 }
 
-#score {
-    color: wheat;
-    text-align: center;
-}
-
 #final-score {
     font-family: "VT323", monospace;
     color: antiquewhite;
     font-size: 3rem;
-}
-
-#instructions {
-    font-family: "VT323", monospace;
-    display: flex;
 }
